report_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai==0.3.2
python-dotenv
